avatar module added and form module updated
