+++ js/form.js
@@ -1,6 +1,7 @@
 import { addClass, removeClass, setItemsAttribute, removeItemsAttribute, addPopup } from './util.js';
 import { setMainPinDefault, closeMapPopup } from './map.js';
 import { sendData } from './server.js';
+import { removeAvatarPreviews, enableAvatarPreviews } from './avatar.js';
 
 /**
  * Function that disable page's forms and inputs
@@ -38,10 +39,6 @@ const validateAdForm = () => {
   const timeOutInput = adForm.querySelector('#timeout');
   const submitButton = adForm.querySelector('.ad-form__submit');
   const resetButton = adForm.querySelector('.ad-form__reset');
-  const photoAvatarChoser = adForm.querySelector('.ad-form-header__input');
-  const photoAvatarPreview = adForm.querySelector('.ad-form-header__preview');
-  const photoHousingChoser = adForm.querySelector('.ad-form__input');
-  const photoHousingPreview = adForm.querySelector('.ad-form__photo');
   /**
    * Data variables
    */
@@ -60,7 +57,6 @@ const validateAdForm = () => {
   };
   const SUBMIT_BUTTON_BLOCKED_TXT = 'Публикую...';
   const SUBMIT_BUTTON_TXT = 'Опубликовать';
-  const IMAGE_TYPES = ['gif', 'jpg', 'jpeg', 'png'];
 
   /**
    * Prisitne library setup
@@ -162,47 +158,20 @@ const validateAdForm = () => {
   };
 
   /**
-   * Function that set photo to preview element
-   * @param {*} choser - input element for choose image
-   * @param {*} preview - container element for preview
-   */
-  const setPhotoPreview = (choser, preview) => {
-    const file = choser.files[0];
-    const fileName = file.name.toLowerCase();
-
-    const matches = IMAGE_TYPES.some((format) => fileName.endsWith(format));
-    if (matches) {
-      const image = document.createElement('div');
-      if (preview.querySelector('img')) {
-        image.classList.add('preview-container', 'preview-container_small');
-      } else {
-        image.classList.add('preview-container');
-      }
-      image.style.backgroundImage = `url(${URL.createObjectURL(file)})`;
-      preview.append(image);
-    }
-  };
-
-  /**
-   * Function that remove previews images
-   */
-  const removePhotoPreviews = () => {
-    const photoPreviews = adForm.querySelectorAll('.preview-container ');
-    if (photoPreviews) {
-      photoPreviews.forEach((item) => item.remove());
-    }
-  };
-
-  /**
    * Function that reset the form to default values
    */
   const resetForm = () => {
     adForm.reset();
     priceSlider.noUiSlider.set(TYPES_LIST.flat);
     setMainPinDefault();
-    removePhotoPreviews();
+    removeAvatarPreviews();
     closeMapPopup();
   };
+
+  /**
+   * Enable form avatar photo previews
+   */
+  enableAvatarPreviews();
 
   /**
    * Adform event listeners
@@ -224,14 +193,6 @@ const validateAdForm = () => {
 
   priceInput.addEventListener('change', function () {
     priceSlider.noUiSlider.set(this.value);
-  });
-
-  photoAvatarChoser.addEventListener('change', (e) => {
-    setPhotoPreview(e.target, photoAvatarPreview);
-  });
-
-  photoHousingChoser.addEventListener('change', (e) => {
-    setPhotoPreview(e.target, photoHousingPreview);
   });
 
   adForm.addEventListener('submit', (evt) => {
